pyproject.toml
[project]
name = "certifexpert"
version = "0.1.0"
description = "CertifExpert — aviation Type Certification expertise platform (EASA Part 21)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6",
    "python-dotenv>=1.0",
    "lxml>=6.0.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["backend/tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.uv]
package = false
